async load of nk-data
